fix register.ts in slack
