+++ packages/slack/src/services/register.ts
@@ -1,11 +1,21 @@
+import { parse } from '../utils/slash-command-parser';
+
 export class RegisterService {
 
-  // constructor(crowi) {
-  //   this.crowi = crowi;
-  // }
+  receiveBody(parseBody) {
+    const body = parse(parseBody);
+    const commandType = body.growiCommandArgs;
+    const commandArgs = body.growiCommandType;
+
+    if (commandType === 'register') {
+      this.openModal();
+      return;
+    }
+  }
 
   openModal() {
-    console.log('waimatu');
+    console.log('open');
   }
+
 
 }
